feat: implement Race class and refactor RaceController
Implement Race class to encapsulate race logic, including round execution, progress display, and winner announcement. Refactor RaceController to utilize the new Race class for managing race execution.
